feat: 🎸 create reusable <JsonCrdtConstant> component
