+++ src/ClickableJsonCrdt/nodes/JsonCrdtConNode.tsx
@@ -1,11 +1,9 @@
 import * as React from 'react';
 import {JsonCrdtProperty} from '../JsonCrdtProperty';
 import {JsonCrdtRegion} from '../JsonCrdtRegion';
-import {JsonAtom} from '../../JsonAtom';
 import {useStyles} from '../../context/style';
-import {ClickableJson} from '../../ClickableJson';
 import {id} from '../utils';
-import {useFocus} from '../../context/focus';
+import {JsonCrdtConstant} from '../JsonCrdtConstant';
 import type {ConNode} from 'json-joy/lib/json-crdt';
 import type {NodeRef} from '../NodeRef';
 
@@ -15,36 +13,11 @@ export interface JsonCrdtConNodeProps {
 
 export const JsonCrdtConNode: React.FC<JsonCrdtConNodeProps> = ({node}) => {
   const {formal} = useStyles();
-  const {focus} = useFocus();
-  const [viewJson, setViewJson] = React.useState(false);
-
-  const view = node.node.view();
-
-  const handleAtomClick = () => {
-    if (view && typeof view === 'object') {
-      setViewJson(!viewJson);
-    }
-  };
 
   return (
     <JsonCrdtRegion node={node}>
       <JsonCrdtProperty node={node} />
-      {viewJson ? (
-        <span style={{display: 'inline-block', verticalAlign: 'top', margin: '-1px'}}>
-          <ClickableJson
-            readonly
-            compact
-            noCollapseToggles
-            pfx={id(node)}
-            doc={node.node.view()}
-            onFocus={(p) => {
-              if (p !== null) focus(id(node));
-            }}
-          />
-        </span>
-      ) : (
-        <JsonAtom value={node.node.view()} onClick={handleAtomClick} />
-      )}
+        <JsonCrdtConstant id={id(node)} view={node.node.view()} />
       {!!formal && ','}
     </JsonCrdtRegion>
   );
